+++ src/WebJobs.Script/Workers/Rpc/RpcWorkerConstants.cs
@@ -48,6 +48,8 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Rpc
         public const string UseNullableValueDictionaryForHttp = "UseNullableValueDictionaryForHttp";
         public const string SharedMemoryDataTransfer = "SharedMemoryDataTransfer";
         public const string FunctionDataCache = "FunctionDataCache";
+        public const string AcceptsListOfFunctionLoadRequests = "AcceptsListOfFunctionLoadRequests";
+        public const string EnableUserCodeException = "EnableUserCodeException";
         public const string SupportsLoadResponseCollection = "SupportsLoadResponseCollection";
 
         // Host Capabilities